add query fixed

+++ db/queries.js
@@ -7,8 +7,8 @@ FROM topics ;`;
 
 exports.querySelectArticles =
 `SELECT articles.article_id, articles.title, articles.topic, articles.author, 
-articles.created_at, articles.votes, articles.article_img_url,
-COUNT(comments.comment_id) AS comment_count
+articles.created_at, articles.votes::integer, articles.article_img_url,
+COUNT(comments.comment_id)::integer AS comment_count
 FROM articles 
 LEFT JOIN comments ON 
 articles.article_id = comments.article_id
